+++ news.md
@@ -3,6 +3,7 @@ New in v2.5
 - Emacs 24.4 compatible.
 - Integration with [Projectile](https://github.com/bbatsov/projectile)!
 - Display `current-directory` in Shell and eshell.
+- New value for `sml/theme`: `automatic` (highly recommended).
 - `sml/apply-theme` is interactive and has completion.
 - Smart-mode-line themes are now regular themes, which makes it easier to define new ones. You can activate them with `(load-theme 'smart-mode-line-light)`, though `(sml/apply-theme 'light)` still works.